add the QWERTY keymap to the config directory for use with GUI keymappers, and update the converstion script to go both ways
